tests/web_platform_tests/graph/personal-graph-basic.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>PersonalGraph Basic Operations</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

// PersonalGraph API — Web Platform Tests
// Tests navigator.graph.create(), addTriple(), queryTriples(), etc.

promise_test(async t => {
  assert_true('graph' in navigator, 'navigator.graph exists');
  assert_true(typeof navigator.graph.create === 'function');
  assert_true(typeof navigator.graph.list === 'function');
  assert_true(typeof navigator.graph.get === 'function');
  assert_true(typeof navigator.graph.remove === 'function');
}, 'navigator.graph interface exists with expected methods');

promise_test(async t => {
  const graph = await navigator.graph.create('Test Graph');
  assert_true(graph instanceof PersonalGraph, 'returns PersonalGraph');
  assert_true(graph.uuid.length > 0, 'has UUID');
  assert_equals(graph.name, 'Test Graph');
  assert_equals(graph.state, 'private');

  // Cleanup
  await navigator.graph.remove(graph.uuid);
}, 'create() returns a PersonalGraph with correct properties');

promise_test(async t => {
  const g1 = await navigator.graph.create('Graph 1');
  const g2 = await navigator.graph.create('Graph 2');

  const list = await navigator.graph.list();
  assert_true(list.length >= 2, 'list contains created graphs');

  const uuids = list.map(g => g.uuid);
  assert_true(uuids.includes(g1.uuid));
  assert_true(uuids.includes(g2.uuid));

  await navigator.graph.remove(g1.uuid);
  await navigator.graph.remove(g2.uuid);
}, 'list() returns all created graphs');

promise_test(async t => {
  const graph = await navigator.graph.create('Lookup');
  const found = await navigator.graph.get(graph.uuid);
  assert_not_equals(found, null);
  assert_equals(found.uuid, graph.uuid);

  const notFound = await navigator.graph.get('nonexistent-uuid');
  assert_equals(notFound, null);

  await navigator.graph.remove(graph.uuid);
}, 'get() retrieves by UUID, returns null for missing');

promise_test(async t => {
  const graph = await navigator.graph.create('ToRemove');
  const result = await navigator.graph.remove(graph.uuid);
  assert_true(result, 'remove returns true');

  const again = await navigator.graph.remove(graph.uuid);
  assert_false(again, 'remove returns false for missing');
}, 'remove() deletes a graph');

promise_test(async t => {
  const graph = await navigator.graph.create('Triples');

  const triple = new SemanticTriple('ex:alice', 'ex:bob', 'foaf:knows');
  const signed = await graph.addTriple(triple);
  assert_true(signed instanceof SignedTriple, 'returns SignedTriple');
  assert_equals(signed.data.source, 'ex:alice');
  assert_equals(signed.data.target, 'ex:bob');
  assert_true(signed.author.startsWith('did:key:'), 'has DID author');
  assert_true(signed.timestamp.length > 0, 'has timestamp');
  assert_true(signed.proof.signature.length > 0, 'has signature');

  await navigator.graph.remove(graph.uuid);
}, 'addTriple() signs and returns a SignedTriple');

promise_test(async t => {
  const graph = await navigator.graph.create('BatchTest');

  const triples = [
    new SemanticTriple('ex:alice', 'ex:bob', 'foaf:knows'),
    new SemanticTriple('ex:alice', 'Alice', 'foaf:name'),
  ];
  const results = await graph.addTriples(triples);
  assert_equals(results.length, 2);

  await navigator.graph.remove(graph.uuid);
}, 'addTriples() batch adds multiple triples');

promise_test(async t => {
  const graph = await navigator.graph.create('QueryTest');

  await graph.addTriple(new SemanticTriple('ex:alice', 'ex:bob', 'foaf:knows'));
  await graph.addTriple(new SemanticTriple('ex:alice', 'Alice', 'foaf:name'));
  await graph.addTriple(new SemanticTriple('ex:bob', 'Bob', 'foaf:name'));

  const results = await graph.queryTriples({ source: 'ex:alice' });
  assert_equals(results.length, 2, 'filters by source');

  const byPred = await graph.queryTriples({ predicate: 'foaf:name' });
  assert_equals(byPred.length, 2, 'filters by predicate');

  const limited = await graph.queryTriples({ source: 'ex:alice', limit: 1 });
  assert_equals(limited.length, 1, 'respects limit');

  await navigator.graph.remove(graph.uuid);
}, 'queryTriples() filters by source, predicate, and limit');

promise_test(async t => {
  const graph = await navigator.graph.create('SparqlTest');

  await graph.addTriple(new SemanticTriple('ex:alice', 'Alice', 'foaf:name'));

  const result = await graph.querySparql(
    'SELECT ?name WHERE { <ex:alice> <foaf:name> ?name }');
  assert_equals(result.type, 'bindings');
  assert_true(result.bindings.length > 0);

  await navigator.graph.remove(graph.uuid);
}, 'querySparql() executes basic SPARQL queries');

promise_test(async t => {
  const graph = await navigator.graph.create('SnapshotTest');

  await graph.addTriple(new SemanticTriple('ex:a', 'ex:b', 'ex:p'));
  await graph.addTriple(new SemanticTriple('ex:c', 'ex:d', 'ex:q'));

  const snapshot = await graph.snapshot();
  assert_equals(snapshot.length, 2, 'snapshot returns all triples');

  await navigator.graph.remove(graph.uuid);
}, 'snapshot() returns all triples');

promise_test(async t => {
  const graph = await navigator.graph.create('RemoveTripleTest');

  const triple = new SemanticTriple('ex:a', 'ex:b', 'ex:p');
  const signed = await graph.addTriple(triple);

  const removed = await graph.removeTriple(signed);
  assert_true(removed, 'removeTriple returns true');

  const remaining = await graph.snapshot();
  assert_equals(remaining.length, 0, 'triple was removed');

  await navigator.graph.remove(graph.uuid);
}, 'removeTriple() removes a signed triple');

promise_test(async t => {
  const graph = await navigator.graph.create('EventTest');

  let addedEvent = null;
  graph.ontripleadded = e => { addedEvent = e; };

  await graph.addTriple(new SemanticTriple('ex:a', 'ex:b', 'ex:p'));

  // Events may be async — give a tick.
  await new Promise(r => setTimeout(r, 100));
  assert_not_equals(addedEvent, null, 'tripleadded event fired');

  await navigator.graph.remove(graph.uuid);
}, 'ontripleadded event fires when triple is added');
</script>
